requirements.txt
Flask
python-dotenv
Flask-Caching
pylti1p3
Werkzeug
requests
PyJWT
cryptography
